Refactor device and environment utilities
