+++ src/fairseq2/recipes/utils/setup.py
@@ -8,23 +8,17 @@ from typing import Dict
 
 import torch
 
-from fairseq2.gang import (
-    Gang,
-    determine_default_device,
-    setup_default_gang,
-    setup_parallel_gangs,
-)
+from fairseq2.device import determine_default_device
+from fairseq2.gang import Gang, setup_default_gang, setup_parallel_gangs
 from fairseq2.logging import LogWriter
 from fairseq2.recipes.utils.log import log_environment_info
 
 
-def setup_gangs(
-    log: LogWriter, *, tp_size: int = 1, monitored: bool = False
-) -> Dict[str, Gang]:
-    """Set up the root, data, and tensor parallel gangs.
+def setup_root_gang(log: LogWriter, *, monitored: bool = False) -> Gang:
+    """Set up the root gang.
 
-    :param tp_size:
-        The size of tensor parallel gangs.
+    :param log:
+        The log to write to.
     :param monitored:
         If ``True``,  puts a monitored barrier before every collective call.
     """
@@ -32,11 +26,31 @@ def setup_gangs(
 
     log_environment_info(log, device)
 
+    # In case we run on Ampere or later, use TF32.
+    torch.set_float32_matmul_precision("high")
+
     log.info("Initializing the root gang.")
 
-    root_gang = setup_default_gang(monitored=monitored)
+    gang = setup_default_gang(monitored=monitored)
 
     log.info("Root gang initialized.")
+
+    return gang
+
+
+def setup_gangs(
+    log: LogWriter, *, tp_size: int = 1, monitored: bool = False
+) -> Dict[str, Gang]:
+    """Set up the root, data, and tensor parallel gangs.
+
+    :param log:
+        The log to write to.
+    :param tp_size:
+        The size of tensor parallel gangs.
+    :param monitored:
+        If ``True``,  puts a monitored barrier before every collective call.
+    """
+    root_gang = setup_root_gang(log, monitored=monitored)
 
     log.info("Initializing data and tensor parallel gangs.")
 
@@ -48,8 +62,5 @@ def setup_gangs(
         ) from ex
 
     log.info("Data and tensor parallel gangs initialized.")
-
-    # In case we run on Ampere or later, use TF32.
-    torch.set_float32_matmul_precision("high")
 
     return gangs
