+++ src/fairseq2/recipes/wav2vec2/asr/train.py
@@ -33,7 +33,7 @@ from fairseq2.optim import AdamW
 from fairseq2.optim.lr_scheduler import TriStageLR
 from fairseq2.recipes.trainer import StandardTrainer
 from fairseq2.recipes.utils.log import log_model
-from fairseq2.recipes.utils.setup import setup_gangs
+from fairseq2.recipes.utils.setup import setup_root_gang
 from fairseq2.recipes.wav2vec2.asr.criterion import Wav2Vec2AsrCriterion
 from fairseq2.typing import CPU, META, DataType
 from fairseq2.utils.profiler import Profiler, Stopwatch
@@ -194,9 +194,7 @@ def load_wav2vec2_asr_trainer(
     """Load a wav2vec 2.0 ASR tainer."""
     wall_watch = Stopwatch(start=True)
 
-    gangs = setup_gangs(log, monitored=config.monitored_gang)
-
-    dp_gang = gangs["dp"]  # data
+    gang = setup_root_gang(log, monitored=config.monitored_gang)
 
     log.info("Loading {} tokenizer.", config.tokenizer_name)
 
@@ -216,7 +214,7 @@ def load_wav2vec2_asr_trainer(
     train_data_reader = dataset.create_reader(
         split=config.train_split,
         tokenizer=tokenizer,
-        gang=dp_gang,
+        gang=gang,
         dtype=config.dtype,
         min_audio_len=config.min_audio_len,
         max_audio_len=config.max_audio_len,
@@ -231,7 +229,7 @@ def load_wav2vec2_asr_trainer(
     valid_data_reader = dataset.create_reader(
         split=config.valid_split,
         tokenizer=tokenizer,
-        gang=dp_gang,
+        gang=gang,
         dtype=config.dtype,
         min_audio_len=config.min_audio_len,
         max_audio_len=config.max_audio_len,
@@ -246,19 +244,19 @@ def load_wav2vec2_asr_trainer(
     log.info("Dataset loaded.")
 
     # Set up the checkpoint manager.
-    if dp_gang.size > 1 and config.data_parallelism == "ddp":
+    if gang.size > 1 and config.data_parallelism == "ddp":
         replicated_keys = ["_model", "_optimizer"]
     else:
         replicated_keys = []
 
     checkpoint_manager = FileCheckpointManager(
         output_dir.joinpath("checkpoints"),
-        dp_gang,
+        gang,
         model_key="_model",
         replicated_keys=replicated_keys,
     )
 
-    rng_bag = RngBag.from_device_defaults(CPU, dp_gang.device)
+    rng_bag = RngBag.from_device_defaults(CPU, gang.device)
 
     # Set the seed for model initialization.
     rng_bag.manual_seed(config.seed)
@@ -274,9 +272,9 @@ def load_wav2vec2_asr_trainer(
     if not has_checkpoint:
         log.info("Loading pretrained {} model on rank 0.", config.pretrained_model_name)
 
-        if dp_gang.rank == 0:
+        if gang.rank == 0:
             pt_model = load_wav2vec2_model(
-                config.pretrained_model_name, device=dp_gang.device, dtype=torch.float32
+                config.pretrained_model_name, device=gang.device, dtype=torch.float32
             )
 
             share_parameters(pt_model.encoder_frontend, model.encoder_frontend)
@@ -287,16 +285,16 @@ def load_wav2vec2_asr_trainer(
 
             del pt_model
 
-        dp_gang.barrier()
+        gang.barrier()
 
         log.info("Pretrained model loaded on rank 0 and parameters shared with the ASR model.")  # fmt: skip
 
         log.info("Initialize the output linear layer on rank 0.")
 
-        if dp_gang.rank == 0:
-            to_device(model, dp_gang.device)
+        if gang.rank == 0:
+            to_device(model, gang.device)
 
-        dp_gang.barrier()
+        gang.barrier()
 
         log.info("Output linear layer initialized.")
 
@@ -310,8 +308,8 @@ def load_wav2vec2_asr_trainer(
     dp_model: Module
 
     # Set up data parallelism.
-    if dp_gang.size == 1:
-        to_device(model, dp_gang.device)
+    if gang.size == 1:
+        to_device(model, gang.device)
 
         dp_model = model
     else:
@@ -320,7 +318,7 @@ def load_wav2vec2_asr_trainer(
         if config.data_parallelism == "ddp":
             find_unused_params = config.freeze_encoder_for_n_steps > 0
 
-            dp_model = to_ddp(model, dp_gang, find_unused_parameters=find_unused_params)
+            dp_model = to_ddp(model, gang, find_unused_parameters=find_unused_params)
         elif config.data_parallelism == "fsdp":
             if config.freeze_encoder_for_n_steps != 0:
                 raise ValueError(
@@ -338,7 +336,7 @@ def load_wav2vec2_asr_trainer(
 
             dp_model = to_fsdp(
                 model,
-                dp_gang,
+                gang,
                 wrap_policy,
                 ignored_modules=ignored_modules,
                 skip_init=True,
@@ -360,7 +358,7 @@ def load_wav2vec2_asr_trainer(
             model.encoder, dynamic=True, options={"shape_padding": True}
         )
 
-        dp_gang.barrier()
+        gang.barrier()
 
         log.info("Encoder compiled.")
 
@@ -369,7 +367,7 @@ def load_wav2vec2_asr_trainer(
     # Initialize the criterion and the optimizer.
     criterion = Wav2Vec2AsrCriterion(
         dp_model,
-        dp_gang,
+        gang,
         tokenizer,
         freeze_encoder_for_n_steps=config.freeze_encoder_for_n_steps,
     )
@@ -388,16 +386,16 @@ def load_wav2vec2_asr_trainer(
     tb_dir = output_dir.joinpath("tb")
 
     profiler = Profiler(
-        skip_first=15, active=3, log_dir=tb_dir, gang=dp_gang, enabled=config.profile
+        skip_first=15, active=3, log_dir=tb_dir, gang=gang, enabled=config.profile
     )
 
     # Set the seed for training.
-    rng_bag.manual_seed(config.seed + dp_gang.rank)
+    rng_bag.manual_seed(config.seed + gang.rank)
 
     # Initialize the trainer.
     return StandardTrainer[Seq2SeqBatch](
         criterion=criterion,
-        gang=dp_gang,
+        gang=gang,
         dtype=config.dtype,
         optimizer=optimizer,
         lr_scheduler=lr_scheduler,
